Fix Merkle verification, schedule seed derivation, and u32 consistency
Spec compliance fixes (from codex.md audit):

1. Merkle proof verification (§4.4.5):
   - Compare full 32-byte root hash at final level, not truncated 20 bytes
   - Update domain strings to match spec:
     SOLANA_MERKLE_SHREDS_LEAF / SOLANA_MERKLE_SHREDS_NODE

2. Committee schedule seed derivation (§3.3.1):
   - Use spec-compliant seed: SHA256("mcp:committee:" || role || LE64(epoch))
   - Pass role strings ("proposer"/"relay") instead of magic values

3. Consistent u32 for proposer_index across all modules:
   - mcp_replay_reconstruction.rs: SlotReconstructionState, OrderedTransaction
   - mcp_relay_attestation.rs: SlotShredState, ReceivedShredInfo
   - mcp_fee_mechanics.rs: SlotFeeTracker, FeeDeductionResult
   - blockstore.rs: get_all_mcp_shreds_for_slot return type
   - replay_stage.rs: implied_proposers type

All 116 MCP tests pass (59 core + 57 ledger).

+++ core/src/mcp_fee_mechanics.rs
@@ -117,7 +117,7 @@ pub enum FeeDeductionResult {
     /// Fees successfully deducted
     Success {
         fees: TransactionFees,
-        including_proposer: u8,
+        including_proposer: u32,
     },
     /// Fee payer couldn't cover the fees
     InsufficientFunds {
@@ -134,7 +134,7 @@ pub struct SlotFeeTracker {
     /// Slot number
     pub slot: Slot,
     /// Fees accumulated per proposer
-    pub proposer_fees: HashMap<u8, u64>,
+    pub proposer_fees: HashMap<u32, u64>,
     /// Total validator fees accumulated
     pub validator_fees: u64,
     /// Number of transactions processed
@@ -178,7 +178,7 @@ impl SlotFeeTracker {
     }
 
     /// Get fees for a specific proposer
-    pub fn get_proposer_fee(&self, proposer_id: u8) -> u64 {
+    pub fn get_proposer_fee(&self, proposer_id: u32) -> u64 {
         self.proposer_fees.get(&proposer_id).copied().unwrap_or(0)
     }
 }
@@ -199,7 +199,7 @@ pub struct TwoPhaseProcessor {
     /// Fee tracker for the current slot
     fee_tracker: SlotFeeTracker,
     /// Map of txid -> proposer_id for fee routing
-    tx_proposer_map: HashMap<Hash, u8>,
+    tx_proposer_map: HashMap<Hash, u32>,
 }
 
 impl TwoPhaseProcessor {
@@ -223,7 +223,7 @@ impl TwoPhaseProcessor {
     }
 
     /// Get the including proposer for a transaction
-    pub fn get_including_proposer(&self, txid: &Hash) -> Option<u8> {
+    pub fn get_including_proposer(&self, txid: &Hash) -> Option<u32> {
         self.tx_proposer_map.get(txid).copied()
     }
